+++ merlin/display.py
@@ -75,6 +75,7 @@ def display_config_info():
     try:
         conf["broker server"] = broker.get_connection_string(include_password=False)
         sconf["broker server"] = broker.get_connection_string()
+        conf["broker ssl"] = broker.get_ssl_config()
     except ValueError:
         conf["broker server"] = "No broker server configured."